requirements.txt
pillow==11.2.1
PyMuPDF==1.26.0
